requirements.txt
python-dotenv
openai==1.30.1
loguru==0.7.2
torch==2.3.1
torchaudio==2.3.1
numpy==2.2.1
requests==2.32.3
soundfile==0.13.0
gradio==5.13.0
httpx==0.27.2